backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
opencv-python==4.8.0.74
Pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.1.0
